Added Socket files to convert to object-oriented style
