+++ src/error.cpp
@@ -1,0 +1,7 @@
+#include "webserv.hpp"
+
+void	error_handle(std::string type)
+{
+	perror(type.c_str());
+	exit(0);
+}
